Redirect asset request to replacement if asset has one
This implements this bit of behaviour [1] in Whitehall's
AttachmentsController#fail method.

Note that I plan to address the call to `expires_headers` in a
subsequent commit.

Note also that I've implemented this for both Whitehall and Mainstream
assets for completeness, even though the immediate requirement is only
to do this for Whitehall assets.

[1]:
https://github.com/alphagov/whitehall/blob/d32f0b1a3fa331c32dd92bd61174900026a0e41d/app/controllers/attachments_controller.rb#L45-L47