ApplicationController#set_expiry now accepts CacheControlConfiguration
Previously when ApplicationController#set_expiry was called from within
WhitehallMediaController#download, it was using the `max_age` from
AssetManager.whitehall_cache_control, but the other options (i.e.
`public`) from AssetManager.cache_control. This wasn't what was intended
and meant that the AssetManager.whitehall_cache_control options were
being ignored.

As it happens the options for both were the same, so this change doesn't
result in any change in behaviour, but hopefully it's a bit easier to
understand.

I've added a new method, CacheControlConfiguration#expires_in, to make
it possible to override the `max_age` option in the places where it's
necessary.

Note that I've also changed the max_age in AssetsController#show from
`0` to `0.minutes` to make the intention of the call to `expires_in`
clearer. Again this shouldn't cause any change in behaviour.