Set Cache-Control header for replacement redirect
The header now set in WhitehallMediaController#download is intended to
match that set in Whitehall's AttachmentsController#fail method [1] for
anonymous users. I haven't addressed the case where the user is signed
in, because this is a more general problem which I've captured in this
GitHub issue [2].

I've done something similar in MediaController#download, but using the
default expiry for Mainstream assets for consistency. Hopefully we can
bring these two values into line in the not too distant future.

[1]:
https://github.com/alphagov/whitehall/blob/446e60b80fee198540e5abe3a7077a5a7f5f63e5/app/controllers/attachments_controller.rb#L46
[2]: https://github.com/alphagov/asset-manager/issues/494