Register MIME types for extensions used in Whitehall Attachments
We are planning to change nginx so that requests for Whitehall
attachments[1] are served from Asset Manager instead of
Whitehall. We've looked at all the existing attachments under
`attachment_data` and requested examples of them to see what
Content-Type header they are served with[2]. This commit registers the
missing types so that the assets are served with the correct header.

I haven't registered the `.ics` extension because I couldn't find an
example of where Whitehall would serve the asset. The default
`text/calendar` which asset manager uses seems sensible.

[1] Those with URLs that start https://www.gov.uk/government/uploads/system/uploads/attachment_data/
[2] #419 (comment)