Register additional unused Whitehall mime types
These two mime types are registered in the Whitehall application[1]
but no assets exist under the `attachment_data` with these
extensions. However there's no disadvantage to adding them here in
case they are used in the future.

Whitehall also explicitly registers types for `.jpg`, `.jpeg`, and
`.svg` but these match the default ones returned by
`Mime::Type.lookup_by_extension` so I'm not going to register them
here.

Whitehall also registers a type for `.html` but as that is also unused
as an attachment and potentially confusing, I'm not going to register
it in Asset Manager at this time.

[1] https://github.com/alphagov/whitehall/blob/446e60b80fee198540e5abe3a7077a5a7f5f63e5/config/initializers/mime_types.rb