Allow assets to save/send a link to their parents
Whitehall attachments can include a `Link` header containing the URL of
their parent document (added to Whitehall in
https://github.com/alphagov/whitehall/pull/2202). We need to add support
for this functionality to Asset Manager so that we can serve Whitehall
attachments from this app.

Although this is primarily for Whitehall attachments I felt that it'd be
less confusing to add support for it to both non-Whitehall and Whitehall
assets.