Prevent transition from published to draft if replaced
We probably don't ever want to make a published asset draft, because
it's URL will be out in the wild. However, this is definitely the case
when the asset has been replaced, because we'll always want the redirect
to the replacement to be publicly accessible.

This commit adds validation on Asset to prevent such a draft status
transition for replaced assets.