Allow access_limited to be reset on update
While investigating issue 450[1], I noticed that resetting/removing the
access_limited status of an asset wasn't working as expected.

This is because when we pass an empty `access_limited` array to
gds-api-adapters/rest-client it's turned into an empty string in Asset
Manager. This results in an `UnpermittedParameter` exception for the
`access_limited` parameter in Asset Manager as it's expecting an Array
but receives a String. The fix is to detect the empty string in the
controller and convert it back into the empty array that's intended. I
don't think this is particularly nice but I think it's good enough for
now.

Note that I haven't had to use `handle_empty_access_limited_param` in
`WhitehallAssetsController#asset_params` as the problem only occurs when
_updating_ an access-limited asset, and the `WhitehallAssetsController`
is only responsible for _creating_ assets.

I also considered using two fields to implement the access-limiting
functionality: a boolean to control whether access-limiting was enabled
in conjunction with an array of user UIDs who are authorised to view the
asset. I decided against that approach because it's more work and
because it means the implementation in Asset Manager diverges from the
access-limiting functionality in the Content Store.

[1]: https://github.com/alphagov/asset-manager/issues/450