Avoid caching responses made to draft-assets host
Draft assets can be optionally access-limited to a subset of users.
Caching the response to a request for an access-limited draft asset
might mean that unauthorised users can see things they shouldn't be able
to.

This is similar to the behaviour in Whitehall where attachment requests
made by signed-in users aren't cached[1].

[1]: https://github.com/alphagov/whitehall/blob/cc57fbae96042952584d90583952d077be1c59a8/app/controllers/base_attachments_controller.rb#L35