Standardise on a 24 hour cache
This removes one of the differences in behaviour between Whitehall and
mainstream assets.

We're about to start serving in the region of an additional 500,000
assets from Asset Manager and so we've standardised on 24 hours (instead
of 4 hours) in order to reduce the frequency at which caches have to
check for freshness. We've added some notes to
https://github.com/alphagov/asset-manager/issues/518 to give an idea of
the number of requests we might expect to serve.