Add temporary Rake task for creating Whitehall Attachments
In issue #525 we've identified a number of Whitehall attachments that
don't have a corresponding file on the filesystem. A subset of these
attachments have a `replaced_by` object which results in requests for
the original URL being redirected to the `replaced_by` URL. This Rake
task creates assets for these attachments so that the redirects continue
to work once we start serving the attachments from Asset Manager.

The scripts we used to come up with the hardcoded list of redirects are
linked to from issue #525.