Remove non-existent replacement from Rake task
The replacement `AttachmentData` identified by ID 303346 doesn't have a
file on disk, and doesn't have a further replacement. As such, requests
for the path of `AttachmentData` with ID 28419 are redirected to ID
303346 which then returns a 404.

This change means that once we're serving Whitehall attachments from
Asset Manager that a request for the path containing 28419 will
immediately 404 rather than redirecting to the path containing 303346
before returning a 404.