Avoid redirecting to non-accessible assets
It's possible to replace a public asset with a draft asset. Normal
replacement behaviour is to redirect from the old to the new asset; but
we shouldn't redirect requests for the public asset in this instance as
it'll mean we send the user to an asset they don't have permission to
view. Note that we do redirect if the user requested the public asset
from the draft-assets host as it means they're authenticated and
therefore authorised (barring any additional access-limiting) to view
draft assets.

This change follows a similar problem in Whitehall where we were
inadvertently redirecting a subset of Attachments to a non-accessible
draft version. See https://github.com/alphagov/whitehall/pull/3894 for
some more information.