Assets should be valid if their replacement has been deleted
We saw some exceptions when updating the metadata of Whitehall
attachments stored in Asset Manager. Some of these attachments had been
replaced by assets that had subsequently been deleted. This left the
assets in an invalid state and meant that we couldn't update them.

We considered removing the `replacement_id` foreign key when a
replacement is deleted but weren't keen on losing that data. Allowing an
asset to be valid if it's been replaced by a deleted asset feels
slightly safer.

See https://github.com/alphagov/asset-manager/issues/533 for some more
information.