Allow deleted assets to be retrieved using the API
We're seeing a number of problems in Whitehall when it's trying to
update attachments in Asset Manager. A large number of these are being
caused by us not being able to retrieve information about assets that
have been deleted. This change should help improve the situation.

More generally, I think this is a reasonable change to make given that
there's an API restore endpoint that undeletes a previously deleted
asset. I think it's probably useful to be able to query the API to
determine the state of the asset before attempting to delete it.

Note that it's not possible to update a previously deleted asset, nor is
it possible to request the content of a deleted asset.