Ensure we don't get redirect chains with replaced assets
When requesting an asset which has been replaced, the user is
redirected to the replacement.  However, if this replacement has been
replaced, the user gets redirected again.  This can result in a long
chain of redirects before the user finally reaches the current version
of the asset.

This commit copies a feature of whitehall, where older assets are
updated to point to the newest replacement.