Add a second web worker which runs on a different port
This will be used to run a second set of unicorn workers
responsible for handling uploads only. The configuration to proxy
uploads to this second instance will be handled by Nginx in Puppet.
This means that in a development environment, only the single `web`
worker will be used to handle both uploads and reads.