Add auth_bypass_ids fields to Asset model
This is field is a set of ids that are set by a publishing application
which can be used to provide access to an asset on the draft stack
without the user having to log in.

There is a feature in authenticating proxy that allows authentication to
bypassed with a token (introduced in
https://github.com/alphagov/authenticating-proxy/pull/20) and a user can
view a page on draft GOV.UK - however this does not provide access to
assets thus this field is to allow this data to be stored with assets as
well.