Bypass draft authentication for appropriate token
This allows the presence of a token, in either query string or a cookie,
to bypass the authentication required to view a draft asset. This is to
match similar functionality provided in
https://github.com/alphagov/authenticating-proxy which allows making
requests to the draft content store without authentication and is
intended to fill the gap in functionality where an unauthenticated user
can view a page but not that pages assets.

The primary means the token is meant to be provided to asset manager is
through a token within a cookie. This is because it seems an unrealistic
expectation for an app to convert all asset urls to have the token in
the query string, as per accessing content. However assets can still
be accessed by having the token in the query string.

Testing this was quite challenging because of the lack of realistic
warden object in a controller test and the lack of an easy way to pass
in cookies for a request test. Thus there are tests in both of these
which test basically the same thing but at different granularities,
there are more calls to mock users and such as we not rely on a warden
object to exist, and there are now login and not_logged_in methods as
the login ones can share state across tests :-(