Include query params in draft asset redirect
In order to support anonymous preview of draft assets, we want to
include a 'token' query param to be intercepted by the auth proxy.
Unfortunately this is removed by the redirect to the draft instance
of Asset Manager, which breaks the planned functionality.

This modifies the redirect in the MediaController to preserve any query
params that were included with the original request.