Remove request test for positive virus scan
This test no longer runs successfully on GDS issued laptops as the ESET
software detects the contents as a virus and quarantines the file, which
is very hard to debug.

This test is also rather strange as it will still pass even if the file
is not identified as having a virus. This is because the endpoint will
return a 404 until a file has been uploaded to cloud storage.