Fix issue when viewing access limited asset
This fixes an issue when viewing an asset that is access limited
and the user is logged in and supplies a bypass id. The previous logic
would not check the bypass id if the user was logged in and
unsuccessfully try to re-authenticate the user. This fixes the logic to
first check if a user has a correct bypass id, then if the authenticate
the user and then check if the user is permitted in the access limits.