Reduce caching period for assets to 30 minutes
We recently had an issue where an asset was correctly unpublished, but it
took manual intervention to remove it from Fastly cache.  Reducing the period
that an asset is cached to 30 minutes should mitigate this issue somewhat.

The intention is to reduce this further, but we want to mitigate the risk
of hammering asset-manager too much until we've had sufficient time to test
this properly.