Fix RSpec/AnyInstance issues
Previously we stubbed a couple of Asset instance methods in order
to simulate a failure when deleting/restoring an Asset, which was
then used to inaccurately return a 422 response, which the caller
could do nothing about (since there request was valid). As part of
fixing the linting issue, this switches to using 'update!', since
it is more accurate to return a 500 for server-oriented errors.