Add content_type field to Asset
This adds a field that can be used to store the content_type of a file,
this then renames the existing `#content_type` method to be
`#content_type_from_extension` to distinguish it and better explain it's
usage.

The motivation for adding this is that determining content type from
extension is unreliable and increases the surface areas for some attack
vectors. It is common for publishing applications to already determine
the content type of files prior to uploading to Asset Manager so this
seems like more accurate data that can be passed to Asset Manager to be
used.

I debated whether Asset Manager should itself determine content type by
analysing the file contents. However this seemed quite a wide reaching
change that might cause some surprises (performance could be affected
and existing files would need evaluating) and may still not agree with a
publishing app's determination of the content type.

One thing I did also consider and reject was whether Asset Manager
should erase the content_type field when an asset has a file updated but
doesn't change an existing content-type. I concluded it seemed unlikely
that an application would set this once and then not later so it seemed
not something to worry about hugely. Should we decide we need then there
may need to be a little refactoring.