Permit publishing applications to specify a content type
This allows a publishing application to specify the content type of an
asset which overrides Asset Manager's default behaviour of using the
filename as a means to determine an asset's content type.

The problem this resolves is when a publishing application decides an
asset is of one type and then Asset Manager disagrees and serves the
file based on the extension - for example if someone (oddly) uploaded a
PDF with a ".html" extension. This issue could cause a confusing error
experience for users and could potentially lead to security issues
of spoofing types of documents.

In considering the application of this I did wonder whether Asset
Manager should just be changed to determine content type based off
reading the file rather than extension (potentially using the
marcel [1] gem) however I concluded that this would be quite a
substantial change with a tough migration.

I concluded that given publishing applications are already determining
content type manually, validating it and presenting this information
to the Publishing API [2] then they can ultimately be considered the
source of truth for this and it isn't ideal for Asset Manager to do the
same calculation, in case they differ.

This change does however leave some future questions (I'm sorry) about
what to do about content type detection in this repo, it may well be
applicable that one day we switch to automatic content type look up in
Asset Manager and allow publishers to override that.

[1]: https://github.com/basecamp/marcel
[2]: https://github.com/alphagov/content-publisher/blob/153de410e4bc377c4a9f5745fe3304cba626156c/spec/lib/publishing_api_payload/file_attachment_payload_spec.rb#L11