Validate content_type field
This checks the format of this input matches the rules in RFC6838. I
wasn't 100% sure whether this would a good thing to add or not, it felt
wrong to use let arbitrary text be allowed to be persisted in Asset
Manager as a content-type but also seemed something that was a bit of a
pain to validate. I settled on a somewhat loose interpretation of the
RFC rules. This seems to be more relaxed than all media-types I've seen
in the wild (I don't think I've ever seen the type of a media type not
be alpha characters) and I don't think this will present false
positives.

I don't expect there's a need for a parameter to be included with this
(e.g the form of "text/html; charset=UTF-8") should that be the case
a validation of this will be much more complex.