Fix missing content_type headers
This applies a fix for handling the situation where an assets
content_type is nil. In these situations we should fallback to
determining the content type from the files extension.

When a content_type of nil is returned Rails will set this to text/html
and this will cause the response to be misrendered for other content
types.

I'm a bit concerned about the level of mocking in the integration test
for this makes this more of a unit test, however my quick attempts to
unpick this weren't fruitful.