Lowercase http_method before passing it to S3 presigned_url().
aws-sdk-s3 seems to have broken Aws::S3::Object#[presigned_url] when
its http_method argument is not passed as lowercase, as of [v1.94.0].
ActionDispatch uses uppercase HTTP method names and we're passing those
verbatim in media_controller.rb, so we fell foul of this when upgrading
the aws-sdk-s3 gem (despite it being only a minor version upgrade).

    irb(main):017:0> Services.cloud_storage.presigned_url_for(Asset.first(), http_method: 'get')
    [Aws::S3::Client 0 0.002086 0 retries] get_object(bucket:"govuk-assets-integration",key:"46c1f091-5c93-416c-9e46-2cf032ed3734")
    => "https://govuk-assets-integration.s3.eu-west-1.amazonaws.com/46c1f091-5c93-416c-9e46-2cf032ed3734?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIA<redacted>%2F20220504%2Feu-west-1%2Fs3%2Faws4_request&X-Amz-Date=20220504T<redacted>&X-Amz-Expires=60&X-Amz-SignedHeaders=host&X-Amz-Signature=<redacted>"

    irb(main):019:0> Services.cloud_storage.presigned_url_for(Asset.first(), http_method: 'GET')
    Traceback (most recent call last):
            3: from (irb):18
            2: from (irb):19:in `rescue in irb_binding'
            1: from lib/s3_storage.rb:42:in `presigned_url_for'
    ArgumentError (unknown operation "get")

I tried to switch to using ActionDispatch's request_method_symbol()
instead of request_method(), so that we're no longer relying on
undocumented behaviour in S3::presigned_url, but that broke so many
change-detector tests that I gave up on that idea :(

We also don't appear to have any test coverage for the integration with
the S3 API. I'm not going to try to fix that just now though.

[presigned_url]: https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/S3/Object.html#presigned_url-instance_method
[v1.94.0]: https://github.com/aws/aws-sdk-ruby/pull/2512/files#diff-6fb7bb7523c6c6bdcd35ea194d13ed2cb9dd32580a03ddbed8e7c7ac6f9f309bR208