Add S3Storage healthcheck
This tests connectivity to S3 storage via the head_bucket method, which returns OK if the bucket requested exists and the requester has the correct permissions for it. This check is performed as part of the readiness checks, so we will not be able to route traffic to the app if we cannot connect to the bucket.

We return a GovukHealthcheck::CRITICAL status in the following cases:

* an error is raised by the call to head_bucket
* head_bucket returns an unsuccessful response

This is because there is some ambiguity in the documentation vs observed behaviour, so we want to protect against both cases.