Rescue from `StateMachines::InvalidTransition`
During virus scanning, it is possible that the asset has been amended.
In this case, the asset's state will have changed from `unscanned` to
something else.

This subsequent change of state by `scanned_clean` will not be
permitted, as the initial state on this transition must be `unscanned`
[1].

Rescuing from the exception will stop this being reported to Sentry, as
the behaviour is expected.

The comment in the rescue block is required to appease rubocop.

1: https://github.com/alphagov/asset-manager/blob/e56fac303c873bb81287f95f24cd1876b2fe1e16/app/models/asset.rb#L87-L90