Make log and tmp directories consistent in files kept
This changes the way we set-up managing these directories to match
Content Publisher [1].

Rails seems to prefer the convention of .keep over .gitkeep - I don't
have a preference myself.

The reason for this is to improve the handling of tmp because we think
it is being polluted in image creation.

[1]: https://github.com/alphagov/content-publisher/commit/4ba770c5b0e31023a652a41fdd5360c55084011d