Add a temporary task to update the legacy urls for topical event assets
Their model was renamed in Whitehall, and this is used on the fly to generate the asset url, meaning that when a topical event was republished, its asset urls pointed to the new location (/topical_event_featuring_image_data) rather than the legacy url those assets are currently found at (/classification_featuring_image_data). This temporary rake task fixes this issue. A long term fix for this would be to store the image urls in Whitehall, so they are not dynamically generated each time. But this is a pragmatic fix for now.