Fix clamdscan in container image.
Asset Manager worker processes were failing to scan uploaded files
because `clamdscan` didn't exist:

    2022-11-05T15:17:03.216Z 1 TID-at2tl WARN: Errno::ENOENT: No such file or directory - /usr/bin/clamdscan

- Install clamdscan package, which was missing.
- Patch the config files for clamd and freshclam to fix logging and
  permissions.
- Avoid installing the shared-mime-info package twice, as it doesn't
  seem to be needed in the build stage.

Really we want ClamAV to run as a separate service so that we can
sandbox it more restrictively and run it with the absolute minimum of
privileges, but we'll tackle that in another PR.

This change only affects Kubernetes/Docker and not EC2 (i.e. current production).