Do not redirect to draft asset host for internal requests
We are currently redirecting to `draft-assets.publishing.service.gov.uk`
when a draft asset is requested by an internal request made by another
application (i.e. at the hostname of `asset-manager` on our internal
infrastructure).

This means that Frontend (when rendering CSV previews), attempts to
follow the redirect to the draft asset, resulting in an error as it is
not able to authenticate itself at that URL.  Additionally, Frontend
already handles the redirect to the draft host.

Therefore removing this redirect when a draft asset is requested
internally by another application.  Authentication still exists, so the
requesting application must still have a suitable bearer token.