Pin Carrierwave gem at version 2
Bumping carrierwave to version 3.0 caused an error [1] in Sentry. It is
not clear whether any of the breaking changes in v3 have caused this,
nor how long it would take to debug and fix the incompatibilities.

The Whitehall asset management team have pinned carrierwave to v2, and
they have on their roadmap to look at the future of using carrierwave in
whitehall [2].

They've pointed out that whitehall's use of carrierwave is more complex
than asset-manager's, so the issues may be unrelated, but I still think
there's value in coordinating our versions.

[1] Details in the PR that reverted the bump https://github.com/alphagov/asset-manager/pull/1152
[2] https://github.com/alphagov/whitehall/pull/8007