Add logging to help diagnose ongoing issue
We have an ongoing issue where we are seeing the following errors related to
missing files coming from the VirusScanWorker:
- Errno::ENOENT: No such file or directory
- StateMachines::InvalidTransition: Cannot transition state via :scanned_clean
  from :unscanned (Reason(s): File can't be blank)

Assets should go through the following stages:
1. unscanned => saved to the file system and scanned by ClamAV in a background
     job.
2. scanned  => Saved to AWS.
3. uploaded => removed from the file system.

However, it seems like the file has already been removed during the virus
scanning stage.

This adds logging so that we can trace the history of assets which raise the
above errors to try to diagnose the issue.