Resolve `Rails.application.secrets` deprecation warning
In Rails 7.2, `Rails.application.secrets` is being deprecated.

We can solve this issue by using the `config_for` method, which will
rerieve the relevant secret from the `config/secrets.yml` file.

This differs from the method used in other applications (example:
https://github.com/alphagov/transition/pull/1439/commits/2d769bddc038f426f90694d2182b04c34c2af691)
since we require the secret to be set before `config/application.rb` is
executed. This occurs before intializers are run, so the alternative
method is not suitable here.