Read metadata from filesystem only when the file has changed
In production we see errors that file reading is attempted by the `store_metadata`
function when the file is not available.

It looks like there might be a concurrency issue where one process has changed the file state
and possibly removed the local file, whilst another process is attempting to save changes
to the asset model and therefore attempts to read metadata from the file.

This change attempts to minimize the attempts to read the file from the filesystem so
that we can avoid these errors.