Only change the asset state after virus scan completes, if the asset is unscanned
In production it looks like there are concurrency issues and the VirusScannerWorker
executes several times for the same file. When several processes attempt to change
the asset state at the same time, it can causes other issues. Specifically any attempt
to read or remove the local file after the state transition can fail.